Refactor testing environment and server setup
- Removed mongodb-memory-server library and switched to using localhost MongoDB for tests.
- Cleared the database after each connection in the test environment to ensure isolation.
- Added 'coverage' folder to ESLint ignore list.
- Used the built-in `http` module provided by Node.js to programmatically control the server in the test environment
- Separated Express app initialization into app.ts and server-related logic into server.ts for cleaner structure.
