+++ package.json
@@ -56,7 +56,6 @@
     "husky": "^9.1.4",
     "jest": "^29.7.0",
     "lint-staged": "^15.2.9",
-    "mongodb-memory-server": "^10.0.0",
     "nodemon": "^3.1.4",
     "prettier": "^3.3.3",
     "supertest": "^7.0.0",
