+++ src/api/v1/user/__test__/user.test.ts
@@ -1,5 +1,5 @@
 import supertest from 'supertest';
-import { app } from '@/server';
+import { app } from '@/app';
 import { success } from '../user.constant';
 import { ErrorTypeEnum, STATUS_CODES, errorMap } from '@/constants';
