+++ src/index.ts
@@ -1,3 +1,14 @@
-import { server } from './server';
+import { connectDB } from './config';
+import { envConstants } from './constants';
+import { setupInitialRolesAndPermissions, setupInitialUserRole } from './script/initialize';
+import { startServer } from './server';
 
-server();
+(async () => {
+  await connectDB();
+  if (!envConstants.INITIAL_SETUP_DONE) {
+    await setupInitialRolesAndPermissions();
+    await setupInitialUserRole();
+  }
+})();
+
+startServer(envConstants.APP_PORT);
