+++ src/utils/test/user.utils.ts
@@ -1,6 +1,6 @@
 import supertest, { Response } from 'supertest';
-import { app } from '../../server';
-import { STATUS_CODES } from '../../constants';
+import { app } from '@/app';
+import { STATUS_CODES } from '@/constants';
 
 export async function retrieveUser(username: string): Promise<Response> {
   return supertest(app).get(`/api/v1/users/${username}`);
